+++ common-api/src/main/java/cn/featherfly/common/lang/function/SerializableDateSupplier.java
@@ -1,8 +1,0 @@
-package cn.featherfly.common.lang.function;
-
-import java.util.Date;
-
-@FunctionalInterface
-public interface SerializableDateSupplier<T extends Date> extends SerializableSupplier<T> {
-
-}