+++ common-api/src/main/java/cn/featherfly/common/lang/function/SerializableLocalDateSupplier.java
@@ -1,8 +1,0 @@
-package cn.featherfly.common.lang.function;
-
-import java.time.LocalDate;
-
-@FunctionalInterface
-public interface SerializableLocalDateSupplier extends SerializableSupplier<LocalDate> {
-
-}