+++ common-api/src/main/java/cn/featherfly/common/lang/function/SerializableLocalDateTimeSupplier.java
@@ -1,8 +1,0 @@
-package cn.featherfly.common.lang.function;
-
-import java.time.LocalDateTime;
-
-@FunctionalInterface
-public interface SerializableLocalDateTimeSupplier extends SerializableSupplier<LocalDateTime> {
-
-}