+++ common-api/src/main/java/cn/featherfly/common/lang/function/SerializableLocalTimeSupplier.java
@@ -1,8 +1,0 @@
-package cn.featherfly.common.lang.function;
-
-import java.time.LocalTime;
-
-@FunctionalInterface
-public interface SerializableLocalTimeSupplier extends SerializableSupplier<LocalTime> {
-
-}