+++ common-api/src/main/java/cn/featherfly/common/lang/function/SerializableStringSupplier.java
@@ -1,9 +1,0 @@
-package cn.featherfly.common.lang.function;
-
-/**
- * The Interface StringSupplier.
- */
-@FunctionalInterface
-public interface SerializableStringSupplier extends SerializableSupplier<String> {
-
-}